dividing and interface

+++ Calculator/ICalculator.cs
@@ -7,6 +7,6 @@ namespace Calculator
         double Subtract(double value1, double value2);
         double Multiply(double value1, double value2);
         double Divide(double value1, double value2);
-        double Divide(int value1, int value2);
+        int Divide(int value1, int value2);
     }
 }
